requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
coloredlogs==15.0.1
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
h11==0.16.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
packaging==25.0
pillow==11.2.1
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
